modify all folder and files

+++ src/services/user.Service.ts
@@ -7,21 +7,50 @@ const createUserIntoDB = async (user: IUser) => {
   if (await createUser.isUserExists(createUser.userId)) {
     throw new Error('User already exists!')
   }
-  const result = await User.create(user)
-  return result
-}
-
-const addOrderToUser = async (user: IUser) => {
-  const result = await User.findOneAndUpdate(user)
+  const userSelect = await User.create(user)
+  const result = User.findOne({ userId: userSelect.userId }).select({
+    _id: 0,
+    password: 0,
+    orders: 0,
+  })
   return result
 }
 
 const getAllUsers = async (): Promise<IUser[]> => {
-  const result = await User.find()
+  const result = await User.find(
+    {},
+    {
+      username: 1,
+      'fullName.firstName': 1,
+      'fullName.lastName': 1,
+      age: 1,
+      email: 1,
+      'address.street': 1,
+      'address.city': 1,
+      'address.country': 1,
+      _id: 0,
+    }
+  )
   return result
 }
-const getSingleUser = async (userId: number): Promise<IUser | null> => {
-  const result = await User.findOne({ userId })
+
+const getSingleUser = async (
+  userId: number,
+  userData: IUser
+): Promise<IUser | null> => {
+  const user = await User.findOne({ userId }).select({
+    _id: 0,
+    password: 0,
+    orders: 0,
+  })
+  if (!user) {
+    return null
+  }
+  const result: IUser = user.toObject()
+  const createUser = new User(userData)
+  if (await createUser.isUserExists(createUser.userId)) {
+    throw new Error('User already exists!')
+  }
   return result
 }
 
@@ -30,25 +59,27 @@ const updateUser = async (
   userData: IUser
 ): Promise<IUser | null> => {
   const filter: FilterQuery<IUser> = { userId }
-
+  const createUser = new User(userData)
+  if (await createUser.isUserExists(createUser.userId)) {
+    throw new Error('User already exists!')
+  }
   const result = await User.findOneAndUpdate(filter, userData, {
     new: true,
     runValidators: true,
   })
-
   return result
 }
 
-const deleteUser = async (userId: number): Promise<IUser | null> => {
+const deleteUser = async (
+  userId: number,
+  user: IUser
+): Promise<IUser | null> => {
   const filter: FilterQuery<IUser> = { userId }
-
+  const createUser = new User(user)
+  if (await createUser.isUserExists(createUser.userId)) {
+    throw new Error('User already exists!')
+  }
   const result = await User.findOneAndDelete(filter)
-
-  return result
-}
-
-const getUserOrders = async (userId: number): Promise<IUser | null> => {
-  const result = await User.findOne({ userId }, { orders: 1, _id: 0 })
   return result
 }
 
@@ -58,6 +89,4 @@ export const UserServices = {
   getSingleUser,
   updateUser,
   deleteUser,
-  getUserOrders,
-  addOrderToUser,
 }
